gui/requirements.txt
matplotlib==2.0.0
numpy==1.12.1
Pillow==5.3.0
PySide2==5.11.2
